+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+x.x.x Release notes (yyyy-MM-dd)
+=============================================================
+
+### API breaking changes
+
+* None.
+
+### Enhancements
+
+* Added method `-linkingObjectsOfClass:forProperty:` to RLMObject to expose inverse
+  relashinships/backlinks.
+
+### Bugfixes
+
+* Throw an exception when passing an array containing a
+  non-RLMObject to -[RLMRealm addObjects:].
+
+
 0.87.0 Release notes (2014-10-21)
 =============================================================