+++ CHANGELOG.md
@@ -14,6 +14,7 @@ x.x.x Release notes (yyyy-MM-dd)
 
 * Throw an exception when passing an array containing a
   non-RLMObject to -[RLMRealm addObjects:].
+* Fix for crash when deleting an object from multiple threads.
 
 
 0.87.0 Release notes (2014-10-21)